Fixed Pathauto prefixing in certain circumstances.
